Move SourceCoordinate to tool package
We need that instead of a source section to identify not yet loaded source locations.
Make charIndex transient. We rely on line/column to identify location.
