+++ src/som/compiler/Parser.java
@@ -78,7 +78,6 @@ import com.oracle.truffle.api.source.SourceSection;
 import som.VM;
 import som.VmSettings;
 import som.compiler.Lexer.Peek;
-import som.compiler.Lexer.SourceCoordinate;
 import som.compiler.MixinBuilder.MixinDefinitionError;
 import som.compiler.Variable.Local;
 import som.interpreter.SNodeFactory;
@@ -107,6 +106,7 @@ import som.interpreter.nodes.specialized.whileloops.WhileInlinedLiteralsNode;
 import som.vm.Symbols;
 import som.vmobjects.SInvokable;
 import som.vmobjects.SSymbol;
+import tools.SourceCoordinate;
 import tools.highlight.Tags;
 import tools.highlight.Tags.ArgumentTag;
 import tools.highlight.Tags.CommentTag;
